requirements.txt
asgiref==3.8.1
cloudinary==1.39.1
Django==4.2.11
django-cloudinary-storage==0.3.0
djangorestframework==3.15.1
pillow==10.3.0
sqlparse==0.4.4
